Improve allocator handling and error checks throughout
- Add robust error handling and cleanup in channel creation
- Use default allocator if none provided in config
- Fix queue allocation size calculation with size_t casts
- Correct reliable channel flag check with parentheses
- Ensure timer never returns 0 as a timestamp
- Minor style and safety improvements

+++ src/libgilberta/queue.c
@@ -15,7 +15,7 @@
 #include "context.h"
 
 glbqueue* glbqueue_init(glbctx_t* ctx, uint32_t element_size, uint32_t capacity) {
-	void* data = ctx->allocator.malloc(sizeof(glbqueue) + capacity * element_size);
+	void* data = ctx->allocator.malloc(sizeof(glbqueue) + (size_t)capacity * (size_t)element_size);
 	if (!data) {
 		ctx->logger(GLB_LOG_ERROR, "[gilberta] Failed to allocate memory for queue");
 		return NULL;
